refactoring all files

+++ server/routes/leave/remove.js
@@ -1,6 +1,10 @@
 const LeaveDelete = require('../../lib/leave/delete')
 
 module.exports = async (req, res) => {
-  await LeaveDelete.deleteLeave(req.params.leaveId)
+  if (!req.session.user) {
+    res.redirect('/')
+  } else if (req.session.user.role === 'Resource Manager') {
+    await LeaveDelete.deleteLeave(req.params.leaveId)
+  }
   res.redirect(`/${req.params.personId}/person`)
 }
